@@ -2,7 +2,7 @@
   "name": "faster-new",
   "displayName": "Quick New File & Folder",
   "description": "Quickly create files and folders",
-  "version": "1.1.18",
+  "version": "1.1.19",
   "publisher": "wenfangdu",
   "icon": "images/icon.png",
   "repository": {
@@ -58,18 +58,20 @@
     "test": "node ./out/test/runTest.js"
   },
   "devDependencies": {
-    "@types/vscode": "^1.0.0",
     "@types/glob": "^7.1.3",
     "@types/mocha": "^8.0.4",
-    "@types/node": "^12.11.7",
-    "eslint": "^7.19.0",
+    "@types/node": "^14.14.34",
+    "@types/vscode": "^1.0.0",
+    "@types/webpack": "^4.41.26",
     "@typescript-eslint/eslint-plugin": "^4.14.1",
     "@typescript-eslint/parser": "^4.14.1",
+    "eslint": "^7.19.0",
     "glob": "^7.1.6",
     "mocha": "^8.2.1",
-    "typescript": "^4.1.3",
-    "vscode-test": "^1.5.0",
     "ts-loader": "^8.0.14",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.2.3",
+    "vscode-test": "^1.5.0",
     "webpack": "^5.19.0",
     "webpack-cli": "^4.4.0"
   }
