@@ -1,11 +1,7 @@
-//@ts-check
+import { Configuration } from 'webpack'
+import * as path from 'path'
 
-'use strict'
-
-const path = require('path')
-
-/**@type {import('webpack').Configuration}*/
-const config = {
+const config: Configuration = {
   target: 'node', // vscode extensions run in a Node.js-context 📖 -> https://webpack.js.org/configuration/node/
   mode: 'none', // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
   entry: './src/extension.ts', // the entry point of this extension, 📖 -> https://webpack.js.org/configuration/entry-context/
@@ -37,4 +33,5 @@ const config = {
     ],
   },
 }
-module.exports = config
+
+export default config
